renamed rush.utils.termutils -> rush.utils.osdetector, as only functions `is_<operating syste>()` left here
